requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Web dashboard
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv>=1.0.0

# AI integrations
anthropic==0.42.0
markdown==3.7

# PDF export
weasyprint==62.3

# Transcripts (optional)
youtube-transcript-api>=0.6.0
